[리팩토링] 1차 리팩토링
