+++ PayStory/src/main/resources/static/main/js/accountBook/public/createPublicCheck.js
